Update: Measure조회 위한 request body form 수정 및 삭제
